requirements.txt
streamlit==1.29.0
plotly-express==0.4.1
pandas
gdown
streamlit_apexjs
altair
